requirements.txt
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
